Unify Arena-related handlers to one
