+++ Mimir.Worker/ActionHandler/ArenaStateHandler.cs
@@ -1,3 +1,4 @@
+using System.Text.RegularExpressions;
 using Bencodex.Types;
 using Lib9c.Models.States;
 using Libplanet.Crypto;
@@ -9,12 +10,12 @@ using Serilog;
 
 namespace Mimir.Worker.ActionHandler;
 
-public class BattleArenaHandler(IStateService stateService, MongoDbService store)
+public class ArenaStateHandler(IStateService stateService, MongoDbService store)
     : BaseActionHandler(
         stateService,
         store,
         "^battle_arena[0-9]*$",
-        Log.ForContext<BattleArenaHandler>())
+        Log.ForContext<ArenaStateHandler>())
 {
     protected override async Task HandleAction(
         long blockIndex,
@@ -25,12 +26,22 @@ public class BattleArenaHandler(IStateService stateService, MongoDbService store
         IClientSessionHandle? session = null,
         CancellationToken stoppingToken = default)
     {
-        var action = new BattleArena();
-        action.LoadPlainValue(actionPlainValue);
-        await ProcessArena(action, session, stoppingToken);
+        if (Regex.IsMatch(actionType, "^battle_arena[0-9]*$"))
+        {
+            var action = new BattleArena();
+            action.LoadPlainValue(actionPlainValue);
+            await ProcessBattleArena(action, session, stoppingToken);   
+        }
+
+        if (Regex.IsMatch(actionType, "^join_arena[0-9]*$"))
+        {
+            var action = new JoinArena();
+            action.LoadPlainValue(actionPlainValue);
+            await ProcessJoinArena(action, session, stoppingToken);   
+        }
     }
 
-    private async Task ProcessArena(
+    private async Task ProcessBattleArena(
         BattleArena battleArena,
         IClientSessionHandle? session = null,
         CancellationToken stoppingToken = default)
@@ -82,6 +93,37 @@ public class BattleArenaHandler(IStateService stateService, MongoDbService store
             battleArena.enemyAvatarAddress,
             battleArena.championshipId,
             battleArena.round,
+            session,
+            stoppingToken);
+    }
+    
+    private async Task ProcessJoinArena(
+        JoinArena joinArena,
+        IClientSessionHandle? session = null,
+        CancellationToken stoppingToken = default)
+    {
+        var arenaScore = await StateGetter.GetArenaScoreAsync(
+            joinArena.avatarAddress,
+            joinArena.championshipId,
+            joinArena.round,
+            stoppingToken);
+        var arenaInfo = await StateGetter.GetArenaInformationAsync(
+            joinArena.avatarAddress,
+            joinArena.championshipId,
+            joinArena.round,
+            stoppingToken);
+        var avatarState = await StateGetter.GetAvatarState(
+            joinArena.avatarAddress,
+            stoppingToken);
+        var simpleAvatarState = SimplifiedAvatarState.FromAvatarState(avatarState);
+        await ArenaCollectionUpdater.UpsertAsync(
+            Store,
+            simpleAvatarState,
+            arenaScore,
+            arenaInfo,
+            joinArena.avatarAddress,
+            joinArena.championshipId,
+            joinArena.round,
             session,
             stoppingToken);
     }
